requirements.txt
arxiv
requests
beautifulsoup4
pandas
nltk
schedule
python-dotenv
openai
pyyaml
python-dateutil